homework week4
