+++ homework-submission/week1/src/App.js
@@ -3,57 +3,64 @@ import logo from './logo.svg';
 import './App.css';
 import Todos from './components/Todos';
 import AddTodo from './components/AddTodo';
-import data from './sources/data.json'
+import data from './sources/data.json';
+import { inject, observer } from 'mobx-react';
+import Store from './Store'
 
+@inject('Store')
+@observer
 class App extends Component {
-  state = {
-    data
-  }
-    handler = e => {
-      let todos = this.state.data;
-      todos[e.target.id -1].done = !todos[e.target.id -1].done;
-      this.setState({
-        todos
-      })
-    }
+    // handler = e => {
+    //   let todos = this.state.data;
+    //   todos[e.target.id -1].done = !todos[e.target.id -1].done;
+    //   this.setState({
+    //     todos
+    //   })
+    // }
 
-    addTodo = e => {
-      e.preventDefault();
-      const {data} = this.state;
-      const newTodo = {
-        id: data.length + 1,
-        description: e.target.description.value,
-        deadline: e.target.deadline.value,
-        done: false
-      };
-      data.push(newTodo)
-      this.setState({
-        data
-      })
-      e.target.description.value = "";
-      e.target.deadline.value = "";
-    }
+    // addTodo = e => {
+    //   e.preventDefault();
+    //   const {data} = this.state;
+    //   const newTodo = {
+    //     id: data.length + 1,
+    //     description: e.target.description.value,
+    //     deadline: e.target.deadline.value,
+    //     done: false
+    //   };
+    //   data.push(newTodo)
+    //   this.setState({
+    //     data
+    //   })
+    //   e.target.description.value = "";
+    //   e.target.deadline.value = "";
+    // }
 
-    removeTodo = (i) => {
-      this.state.data.splice(i, 1);
-      console.log(i)
-      this.setState({
-        data: this.state.data
-      })
-    }
+    // removeTodo = (i) => {
+    //   this.state.data.splice(i, 1);
+    //   console.log(i)
+    //   this.setState({
+    //     data: this.state.data
+    //   })
+    // }
 
     render() {
+      const {
+        handler,
+        addTodo,
+        removeTodo
+      } = this.props.Store;
     return (
       <div className="App">
         <header className="App-header">
           <img src={logo} className="App-logo" alt="logo" />
           <h1 className="App-title">To-Do App</h1>
         </header>
+
         <AddTodo
           addItem = {this.addTodo} 
           />
         <Todos 
-          items = {this.state.data} 
+          // items = {this.state.data} 
           handler = {this.handler}
           remove = {this.removeTodo}
          />
