+++ homework-submission/week1/src/index.js
@@ -3,6 +3,15 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
+import {Provider} from 'mobx-react';
+import Store from './Store';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+
+const Root = (
+    <Provider store={Store}>
+        <App />
+    </Provider>
+)
+
+ReactDOM.render(Root, document.getElementById('root'));
 registerServiceWorker();
